Integrated the DB to handle all the ORDER manage with the System

+++ api/model/constants.go
@@ -79,7 +79,7 @@ var Menu = []Product{
 	Espresso, Americano, Cappuccino, Latte, TuCowsMilk,
 }
 
-func productMapper(ID int) Product {
+func productMapper(ID int64) Product {
 	switch ID {
 	case ProductID_Espresso:
 		return Espresso
